move common remote sources unpacking testing to its own file
