+++ tests/python/pants_test/backend/jvm/tasks/test_unpack_jars.py
@@ -6,7 +6,6 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 import functools
 import os
-import unittest
 from contextlib import contextmanager
 
 from pants.backend.jvm.targets.jar_library import JarLibrary
@@ -35,45 +34,6 @@ class UnpackJarsTest(TaskTestBase):
         proto_jarfile.writestr('a/b/c/{}.txt'.format(name), 'Some text')
         proto_jarfile.writestr('a/b/c/{}.proto'.format(name), 'message Msg {}')
       yield jar_name
-
-  def test_invalid_pattern(self):
-    with self.assertRaises(UnpackJars.InvalidPatternError):
-      UnpackJars.compile_patterns([45])
-
-  @staticmethod
-  def _run_filter(filename, include_patterns=None, exclude_patterns=None):
-    return UnpackJars._file_filter(
-      filename,
-      UnpackJars.compile_patterns(include_patterns or []),
-      UnpackJars.compile_patterns(exclude_patterns or []))
-
-  def test_file_filter(self):
-    # If no patterns are specified, everything goes through
-    self.assertTrue(self._run_filter("foo/bar.java"))
-
-    self.assertTrue(self._run_filter("foo/bar.java", include_patterns=["**/*.java"]))
-    self.assertTrue(self._run_filter("bar.java", include_patterns=["**/*.java"]))
-    self.assertTrue(self._run_filter("bar.java", include_patterns=["**/*.java", "*.java"]))
-    self.assertFalse(self._run_filter("foo/bar.java", exclude_patterns=["**/bar.*"]))
-    self.assertFalse(self._run_filter("foo/bar.java",
-                                      include_patterns=["**/*/java"],
-                                      exclude_patterns=["**/bar.*"]))
-
-    # exclude patterns should be computed before include patterns
-    self.assertFalse(self._run_filter("foo/bar.java",
-                                      include_patterns=["foo/*.java"],
-                                      exclude_patterns=["foo/b*.java"]))
-    self.assertTrue(self._run_filter("foo/bar.java",
-                                     include_patterns=["foo/*.java"],
-                                     exclude_patterns=["foo/x*.java"]))
-
-  @unittest.expectedFailure
-  def test_problematic_cases(self):
-    """These should pass, but don't"""
-    # See https://github.com/twitter/commons/issues/380.  'foo*bar' doesn't match 'foobar'
-    self.assertFalse(self._run_filter("foo/bar.java",
-                                      include_patterns=['foo/*.java'],
-                                      exclude_patterns=['foo/bar*.java']))
 
   def _make_jar_library(self, coord):
     return self.make_target(spec='unpack/jars:foo-jars',
